ai_parts/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
openai>=1.0.0
python-multipart>=0.0.6
llama-index-core>=0.10.60
llama-index-embeddings-jinaai>=0.5.0
llama-index-vector-stores-chroma>=0.2.2
chromadb>=0.5.17
